# docx/deliverables/openapi.yaml
openapi: 3.1.0
info:
  title: 学生生活一站式社区 AI 助手 API
  version: 0.5.0
  description: |
    详细设计增量 1：公共基础与 M4 平台治理；增量 2：M2 校园服务中心；
    增量 3：M1 AI 知识与对话；增量 4：M3 校园社区与互助；
    增量 5：M5 多智能体、Tool Runtime、小模型工程、训练评估与 M2 演示电费。
    已发布 operationId 不得改变。
  contact:
    name: 项目开发小组
  license:
    name: Educational Use Only
    identifier: LicenseRef-Educational-Use-Only
servers:
  - url: http://localhost:8000
    description: 后端本地开发
  - url: http://localhost
    description: Docker Compose 演示环境
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/lost-found:
    get:
      tags: [LostFound]
      operationId: listLostFoundItems
      summary: 分页筛选已发布失物/拾物信息
      description: 普通列表只返回 contact_hint，不返回联系方式密文或完整联系方式。
      x-permissions: [community:read]
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: item_type
          in: query
          schema: {$ref: "#/components/schemas/LostFoundItemType"}
        - name: category
          in: query
          schema: {type: string, maxLength: 50}
        - name: location
          in: query
          schema: {type: string, maxLength: 100}
        - name: occurred_from
          in: query
          schema: {type: string, format: date-time}
        - name: occurred_to
          in: query
          schema: {type: string, format: date-time}
        - name: mine
          in: query
          schema: {type: boolean, default: false}
      responses:
        "200":
          description: 失物/拾物分页
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LostFoundItemPageResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "422": {$ref: "#/components/responses/ValidationError"}
    post:
      tags: [LostFound]
      operationId: createLostFoundItem
      summary: 发布失物或拾物信息
      description: 联系方式由应用层加密后保存；内容先经过 M4 扫描。
      x-permissions: [community:write]
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LostFoundItemCreateRequest"}
      responses:
        "201":
          description: 信息已创建并安排候选匹配
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LostFoundItemResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /api/v1/lost-found/{item_id}:
    get:
      tags: [LostFound]
      operationId: getLostFoundItem
      summary: 获取失物/拾物详情
      x-permissions: [community:read]
      parameters:
        - $ref: "#/components/parameters/LostFoundItemId"
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: 可见详情；完整联系方式默认不返回
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LostFoundItemResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      tags: [LostFound]
      operationId: updateLostFoundItem
      summary: 所有者修改记录并重新匹配/审核
      x-permissions: [community:write]
      x-resource-rules: [owner_or_community_moderator]
      parameters:
        - $ref: "#/components/parameters/LostFoundItemId"
        - $ref: "#/components/parameters/XRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LostFoundItemUpdateRequest"}
      responses:
        "200":
          description: 修改后的记录
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LostFoundItemResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
    delete:
      tags: [LostFound]
      operationId: deleteLostFoundItem
      summary: 所有者或运营员逻辑删除记录
      x-permissions: [community:write]
      x-resource-rules: [owner_or_community_moderator]
      parameters:
        - $ref: "#/components/parameters/LostFoundItemId"
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "204": {description: 记录和候选关系已失效}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /api/v1/lost-found/{item_id}/claims:
    post:
      tags: [LostFound]
      operationId: createLostFoundClaim
      summary: 提交关键特征验证请求
      description: 目标记录联系方式仍保持隐藏；验证材料加密保存。
      x-permissions: [community:write]
      parameters:
        - $ref: "#/components/parameters/LostFoundItemId"
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LostFoundClaimCreateRequest"}
      responses:
        "201":
          description: 认领请求已创建
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LostFoundClaimResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /api/v1/lost-found-claims:
    get:
      tags: [LostFound]
      operationId: listMyLostFoundClaims
      summary: 查询本人发起或需要本人处理的认领请求
      x-permissions: [community:write]
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: role
          in: query
          schema: {type: string, enum: [claimant, owner, all], default: all}
        - name: status
          in: query
          schema: {$ref: "#/components/schemas/LostFoundClaimStatus"}
      responses:
        "200":
          description: 本人认领请求分页
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LostFoundClaimPageResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /api/v1/lost-found-claims/{claim_id}:
    get:
      tags: [LostFound]
      operationId: getLostFoundClaim
      summary: 获取本人参与的认领详情
      x-permissions: [community:write]
      x-resource-rules: [claimant_or_target_owner]
      parameters:
        - $ref: "#/components/parameters/LostFoundClaimId"
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: 认领详情；验证材料只向目标所有者和提交者返回
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LostFoundClaimResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/lost-found-claims/{claim_id}/decision:
    post:
      tags: [LostFound]
      operationId: decideLostFoundClaim
      summary: 目标记录所有者验证或拒绝关键特征
      x-permissions: [community:write]
      x-resource-rules: [target_owner_only]
      parameters:
        - $ref: "#/components/parameters/LostFoundClaimId"
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LostFoundClaimDecisionRequest"}
      responses:
        "200":
          description: 决策后的认领请求
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LostFoundClaimResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /api/v1/lost-found-claims/{claim_id}/contact:
    get:
      tags: [LostFound]
      operationId: getLostFoundClaimContact
      summary: 验证通过后交换双方联系方式
      x-permissions: [community:write]
      x-resource-rules: [verified_claim_participant]
      parameters:
        - $ref: "#/components/parameters/LostFoundClaimId"
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: 双方联系方式；响应禁止缓存并写审计
          headers:
            Cache-Control:
              schema: {type: string, const: no-store}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LostFoundContactResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /api/v1/lost-found-claims/{claim_id}/completion:
    post:
      tags: [LostFound]
      operationId: confirmLostFoundClaimCompletion
      summary: 认领双方分别确认完成
      description: 单方确认保持 verified；双方均确认后 claim 和相关记录进入 completed。
      x-permissions: [community:write]
      x-resource-rules: [verified_claim_participant]
      parameters:
        - $ref: "#/components/parameters/LostFoundClaimId"
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LostFoundCompletionRequest"}
      responses:
        "200":
          description: 最新确认状态
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LostFoundClaimResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
    internalServiceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-token
      description: 仅受信 Agent Runtime 使用的独立服务凭证，不接受用户 JWT
...
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: 客户端可提供；非法或缺失时服务端生成新 ID
      schema: {type: string, minLength: 8, maxLength: 64}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 同一用户和接口 24 小时内重放返回首次结果
      schema: {type: string, minLength: 8, maxLength: 128}
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
    LostFoundItemId:
      name: item_id
      in: path
      required: true
      schema: {type: string, format: uuid}
    LostFoundClaimId:
      name: claim_id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    Unauthorized:
      description: Access Token 缺失、无效或已过期
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
          examples:
            unauthenticated:
              $ref: "#/components/examples/UnauthorizedError"
    Forbidden:
      description: 已认证但缺少操作或资源权限
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
          example:
            code: AUTH_FORBIDDEN
            message: 没有执行该操作的权限
            details: []
            request_id: 01J2ABCDEF0123456789ABCDE0
            timestamp: 2026-07-12T08:00:00Z
    NotFound:
      description: 资源不存在，或当前用户无权感知该资源
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Conflict:
      description: 唯一约束、状态机或乐观锁冲突
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
          example:
            code: RESOURCE_VERSION_CONFLICT
            message: 数据已被其他操作更新，请刷新后重试
            details: []
            request_id: 01J2ABCDEF0123456789ABCDE0
            timestamp: 2026-07-12T08:00:00Z
    ValidationError:
      description: 字段格式或业务输入校验失败
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
          example:
            code: VALIDATION_ERROR
            message: 请求参数校验失败
            details:
              - field: password
                reason: 至少 10 个字符，并包含字母、数字和特殊字符
            request_id: 01J2ABCDEF0123456789ABCDE0
            timestamp: 2026-07-12T08:00:00Z
...
  examples:
    UnauthorizedError:
      value:
        code: AUTH_UNAUTHORIZED
        message: 登录状态无效，请重新登录
        details: []
        request_id: 01J2ABCDEF0123456789ABCDE0
        timestamp: 2026-07-12T08:00:00Z
  schemas:
    ResponseMeta:
      type: object
      required: [code, message, request_id, timestamp]
      properties:
        code: {type: string, example: OK}
        message: {type: string, example: success}
        request_id: {type: string, maxLength: 64}
        timestamp: {type: string, format: date-time}
...
    ErrorDetail:
      type: object
      required: [reason]
      additionalProperties: false
      properties:
        field: {type: [string, "null"]}
        reason: {type: string}
        context:
          type: [object, "null"]
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [code, message, details, request_id, timestamp]
      additionalProperties: false
      properties:
        code: {type: string}
        message: {type: string}
        details:
          type: array
          items: {$ref: "#/components/schemas/ErrorDetail"}
        request_id: {type: string}
        timestamp: {type: string, format: date-time}
    PageMeta:
      type: object
      required: [page, page_size, total, total_pages]
      additionalProperties: false
      properties:
        page: {type: integer, minimum: 1}
        page_size: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
        total_pages: {type: integer, minimum: 0}
...
    PublicAuthor:
      type: object
      required: [display_name, is_anonymous]
      additionalProperties: false
      properties:
        user_id: {type: [string, "null"], format: uuid}
        display_name: {type: string}
        avatar_url: {type: [string, "null"], format: uri}
        is_anonymous: {type: boolean}
...
    LostFoundItemType:
      type: string
      enum: [lost, found]
    LostFoundItemStatus:
      type: string
      enum: [pending_review, published, claiming, completed, closed, rejected, deleted]
    ContactType:
      type: string
      enum: [phone, email, wechat, other]
    LostFoundItem:
      type: object
      required: [id, owner, item_type, title, category, description, occurred_at, location, contact_type, contact_hint, status, version, created_at, updated_at]
      additionalProperties: false
      properties:
        id: {type: string, format: uuid}
        owner: {$ref: "#/components/schemas/PublicAuthor"}
        item_type: {$ref: "#/components/schemas/LostFoundItemType"}
        title: {type: string, maxLength: 120}
        category: {type: string, maxLength: 50}
        description: {type: string, maxLength: 2000}
        occurred_at: {type: string, format: date-time}
        location: {type: string, maxLength: 200}
        contact_type: {$ref: "#/components/schemas/ContactType"}
        contact_hint: {type: string, maxLength: 50}
        status: {$ref: "#/components/schemas/LostFoundItemStatus"}
        moderation_case_id: {type: [string, "null"], format: uuid}
        published_at: {type: [string, "null"], format: date-time}
        completed_at: {type: [string, "null"], format: date-time}
        version: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    LostFoundItemCreateRequest:
      type: object
      required: [item_type, title, category, description, occurred_at, location, contact_type, contact_value]
      additionalProperties: false
      properties:
        item_type: {$ref: "#/components/schemas/LostFoundItemType"}
        title: {type: string, minLength: 2, maxLength: 120}
        category: {type: string, minLength: 1, maxLength: 50}
        description: {type: string, minLength: 5, maxLength: 2000}
        occurred_at: {type: string, format: date-time}
        location: {type: string, minLength: 2, maxLength: 200}
        contact_type: {$ref: "#/components/schemas/ContactType"}
        contact_value: {type: string, minLength: 3, maxLength: 200, writeOnly: true}
    LostFoundItemUpdateRequest:
      type: object
      required: [version]
      additionalProperties: false
      properties:
        title: {type: string, minLength: 2, maxLength: 120}
        category: {type: string, minLength: 1, maxLength: 50}
        description: {type: string, minLength: 5, maxLength: 2000}
        occurred_at: {type: string, format: date-time}
        location: {type: string, minLength: 2, maxLength: 200}
        contact_type: {$ref: "#/components/schemas/ContactType"}
        contact_value: {type: string, minLength: 3, maxLength: 200, writeOnly: true}
        version: {type: integer, minimum: 1}
    LostFoundItemResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseMeta"
        - type: object
          required: [data]
          properties:
            data: {$ref: "#/components/schemas/LostFoundItem"}
    LostFoundItemPageData:
      type: object
      required: [items, pagination]
      additionalProperties: false
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/LostFoundItem"}
        pagination: {$ref: "#/components/schemas/PageMeta"}
    LostFoundItemPageResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseMeta"
        - type: object
          required: [data]
          properties:
            data: {$ref: "#/components/schemas/LostFoundItemPageData"}
...
    LostFoundClaimStatus:
      type: string
      enum: [pending, verified, rejected, cancelled, completed]
    LostFoundClaimCreateRequest:
      type: object
      required: [evidence]
      additionalProperties: false
      properties:
        claimant_item_id: {type: [string, "null"], format: uuid}
        evidence: {type: string, minLength: 5, maxLength: 1000, writeOnly: true}
    LostFoundClaimDecisionRequest:
      type: object
      required: [decision, version]
      additionalProperties: false
      properties:
        decision: {type: string, enum: [verified, rejected]}
        reason: {type: [string, "null"], maxLength: 500}
        version: {type: integer, minimum: 1}
    LostFoundCompletionRequest:
      type: object
      required: [version]
      additionalProperties: false
      properties:
        version: {type: integer, minimum: 1}
    LostFoundClaim:
      type: object
      required: [id, target_item, claimant, evidence, status, claimant_confirmed, owner_confirmed, version, created_at, updated_at]
      additionalProperties: false
      properties:
        id: {type: string, format: uuid}
        target_item: {$ref: "#/components/schemas/LostFoundItem"}
        claimant_item_id: {type: [string, "null"], format: uuid}
        claimant: {$ref: "#/components/schemas/PublicAuthor"}
        evidence: {type: string, maxLength: 1000}
        status: {$ref: "#/components/schemas/LostFoundClaimStatus"}
        decision_reason: {type: [string, "null"]}
        claimant_confirmed: {type: boolean}
        owner_confirmed: {type: boolean}
        completed_at: {type: [string, "null"], format: date-time}
        version: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    LostFoundClaimResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseMeta"
        - type: object
          required: [data]
          properties:
            data: {$ref: "#/components/schemas/LostFoundClaim"}
    LostFoundClaimPageData:
      type: object
      required: [items, pagination]
      additionalProperties: false
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/LostFoundClaim"}
        pagination: {$ref: "#/components/schemas/PageMeta"}
    LostFoundClaimPageResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseMeta"
        - type: object
          required: [data]
          properties:
            data: {$ref: "#/components/schemas/LostFoundClaimPageData"}
    ContactParty:
      type: object
      required: [user, contact_type, contact_value]
      additionalProperties: false
      properties:
        user: {$ref: "#/components/schemas/PublicAuthor"}
        contact_type: {$ref: "#/components/schemas/ContactType"}
        contact_value: {type: string}
    LostFoundContactData:
      type: object
      required: [claim_id, requester, counterpart]
      additionalProperties: false
      properties:
        claim_id: {type: string, format: uuid}
        requester: {$ref: "#/components/schemas/ContactParty"}
        counterpart: {$ref: "#/components/schemas/ContactParty"}
    LostFoundContactResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseMeta"
        - type: object
          required: [data]
          properties:
            data: {$ref: "#/components/schemas/LostFoundContactData"}
...